refactor(web): StatCounterOrText primitive 추출 + Impact/AboutCounters 재사용
BoldProjectDetailImpact 가 AboutCounters 의 StatCounterOrText 헬퍼를 동일하게
복사해서 정의 (주석으로도 본인이 인지). audit (MEDIUM) 의 후속 정리.

- 신규 components/primitives/StatCounterOrText.jsx: 동일 로직 lift. 순수 숫자
  매칭 시 StatCounter 카운트업, 단위 섞이면 plain span
- AboutCounters: 로컬 StatCounterOrText 삭제 + primitive import
- BoldProjectDetailImpact: 로컬 StatCounterOrText 삭제 + primitive import

visual / 동작 변화 없음.

Refs #109 #70
